cloud_functions/create_group/requirements.txt
# Function dependencies, for example:
# package>=version
google-cloud-datastore
flask
